Enter bootloader by press SELECT and SEND together
Bootloader modified to enter in bootloader state when SELECT and SEND
are both pressed when the device is inserted on USB port. After entering
bootloader, the alternative activation by Num_Lock and Scroll_Lock are
set by default as well as the current Slot is changed to '0' and the
flag that indicates an Upload has occurred is set. This configuration is
stored on the last address of EEPROM ( E2END )